refactor:지난주 로또번호의 validate로직을 도메인객체로 옮김
