+++ src/main/java/nextstep/step2/view/InputView.java
@@ -1,6 +1,5 @@
 package nextstep.step2.view;
 
-import java.util.Arrays;
 import java.util.NoSuchElementException;
 import java.util.Scanner;
 
@@ -9,7 +8,6 @@ public class InputView {
 	private static final String INPUT_MESSAGE_LAST_LOTTO = "지난 주 당첨 번호를 입력해 주세요.";
 	private static final String WRONG_INPUT_MESSAGE = "구입금액은 필수입니다.";
 	private static final String WRONG_INPUT_LAST_LOTTO = "지난주 로또 당첨 번호는 필수입니다.";
-	private static final String WRONG_WINNER_NUMBERS = "지난주 로또 당첨번호는 모두 숫자여야 합니다.";
 
 	private Scanner scanner;
 
@@ -38,18 +36,9 @@ public class InputView {
 	protected String getInputLottoNumbers() {
 		try {
 			String winnerNumbers = scanner.nextLine();
-			validateLastWinnerNumbers(winnerNumbers);
 			return winnerNumbers;
 		} catch (NoSuchElementException ex) {
 			throw new IllegalArgumentException(WRONG_INPUT_LAST_LOTTO);
-		}
-	}
-
-	protected void validateLastWinnerNumbers(String winnerNumbers) {
-		String[] numbers = winnerNumbers.replaceAll(" ","").split(",");
-		boolean isDigit = Arrays.stream(numbers).allMatch(oddNum -> oddNum.chars().allMatch(Character::isDigit));
-		if (!isDigit) {
-			throw new IllegalArgumentException(WRONG_WINNER_NUMBERS);
 		}
 	}
 }
