+++ src/test/java/nextstep/step2/view/InputViewTest.java
@@ -42,20 +42,4 @@ public class InputViewTest {
 				.isThrownBy(() -> underTest.getInputLottoNumbers())
 				.withMessage("지난주 로또 당첨 번호는 필수입니다.");
 	}
-
-	@Test
-	@DisplayName("지난주 로또 당첨번호는 숫자여야 한다")
-	public void inputWrongLastLottoNumberTest() {
-		InputView underTest = new InputView(new Scanner("1,2,3,a,b,c"));
-		Assertions.assertThatExceptionOfType(IllegalArgumentException.class)
-				.isThrownBy(() -> underTest.validateLastWinnerNumbers("1,2,3,a,b,c"))
-				.withMessage("지난주 로또 당첨번호는 모두 숫자여야 합니다.");
-	}
-
-	@Test
-	@DisplayName("로또 당첨번호에서 공백을 제거한다.")
-	public void inpuLastLottoNumberTest() {
-		InputView underTest = new InputView(new Scanner("1,2,3,4, 5,  6"));
-		underTest.validateLastWinnerNumbers("1,2,3,4, 5,  6");
-	}
 }
